refactor: move thinking labels to chat session constants
