+++ apps/frontend/src/modules/chat-session/components/ThinkingLevelSelect/index.ts
@@ -1,3 +1,2 @@
-export {THINKING_LEVEL_LABELS, THINKING_LEVELS} from './constants.js';
 export {useThinkingLevel} from './hooks/useThinkingLevel.js';
 export {ThinkingLevelSelect} from './ThinkingLevelSelect.js';
