+++ apps/frontend/src/modules/chat-session/constants.ts
