+++ apps/frontend/src/modules/chat-session/index.ts
@@ -11,6 +11,9 @@ export {
 } from './components/ThinkingLevelSelect/index.js';
 export {TitleBarView} from './components/TitleBar/index.js';
 
+// Constants
+export {THINKING_LEVEL_LABELS, THINKING_LEVELS} from './constants.js';
+
 // Contexts (providers + values)
 export {ChatEventBusProvider} from './contexts/ChatEventBusContext/index.js';
 export {
